Initial commit, adds game.
This is a project that started as an idea for a one-page RPG. It's been
kicking around my computer and head for various forms for a few months,
and I've just now cleaned it up and started seriously looking at
packaging it. Sorry for the lack of history before this.