Make build script moderately more flexible.
In the future, I will have the urge to make this more complex. At that
point, someone should point me at this commit message, where I remind my
future self that this is a hobby project, intended for non-techical
people, and that it doesn't need to be complex - the source file is
markdown anyway! People can just use it as-is!

No one wants to be reading complex bash, least of all you, future me.